requirements.txt
backtrader==1.9.74.123
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
common==0.1.2
cycler==0.10.0
dbschema==1.4.2
gevent==1.4.0
greenlet==0.4.15
html5lib==1.0.1
idna==2.8
kiwisolver==1.1.0
lxml==4.4.1
matplotlib==3.1.1
mechanize==0.4.4
numpy==1.17.2
pandas==0.25.1
pandas-datareader==0.8.1
pg8000==1.13.2
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycparser==2.19
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-mock==1.6.0
scramp==1.1.0
six==1.12.0
tdameritrade==0.0.8
testing.common.database==2.0.3
testing.postgresql==1.3.0
tiingo==0.11.0
urllib3==1.25.3
webencodings==0.5.1
websocket-client==0.56.0
